Add Tappable extension for widgets and implement navigation in login screen
This commit introduces a new extension `Tappable` to `Widget` which provides `onTap` and `onLongPress` functionalities using `InkWell`.

The login screen now utilizes this extension for handling taps on text elements and implements navigation to the register and home screens upon button presses.

Additionally, the `GestureDetector` in `FButton` has been moved to wrap the `Container` for better tap handling.