Refactor food widget and cart functionality
This commit refactors the `FoodWidget` to be a `StatefulWidget` and introduces an animation for the add-to-cart button. It also updates the cart cubit to handle adding existing items to the cart by incrementing their quantity and recalculating the total price. Additionally, the cart screen UI has been updated to include an edit cart option and an empty cart message.

Total price calculation now works for multiple orders of same foodentity.